Update and rename index (3).js to snakeMovement.js
